requirements.txt
fastapi
"uvicorn[standart]"
"pydantic[email]"
pydantic-settings
"sqlalchemy[asyncio]"
asyncio
aiosqlite
bcrypt
jwt